Liniting and refactoring
