+++ .rubocop.yml
@@ -1,0 +1,8 @@
+AllCops:
+  TargetRubyVersion: 2.3
+
+Style/StringLiterals:
+  Enabled: false
+
+Metrics/LineLength:
+  Max: 120
