+++ Rakefile
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "bundler/gem_tasks"
 require 'rake/testtask'
 
@@ -8,4 +10,4 @@ Rake::TestTask.new do |t|
 end
 
 desc "Run all tests..."
-task :default => :test
+task default: :test
