+++ bin/terminal_chess
@@ -1,3 +1,4 @@
 #!/usr/bin/env ruby
+# frozen_string_literal: true
 
 require 'terminal_chess'
