+++ lib/terminal_chess.rb
@@ -1,16 +1,16 @@
 #!/usr/bin/env ruby
+# frozen_string_literal: true
 
-$LOAD_PATH << __FILE__ # '.'
+$LOAD_PATH << './lib'
 
-require_relative 'local_chess_client'
-require_relative 'network_chess_client'
-require_relative 'server'
+require 'terminal_chess/local_chess_client'
+require 'terminal_chess/network_chess_client'
+require 'terminal_chess/server'
 
 if ENV["SERVER"]
-  ChessServer.new
+  TerminalChess::ChessServer.new
 elsif ENV["NGROK"]
-  NetworkChessClient.new(ENV["NGROK"])
+  TerminalChess::NetworkChessClient.new(ENV["NGROK"])
 else
-  LocalChessClient.new
+  TerminalChess::LocalChessClient.new
 end
-
