+++ lib/terminal_chess/version.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module TerminalChess
-  VERSION = "0.2.0"
+  VERSION = "0.2.1"
 end
