+++ terminal_chess.gemspec
@@ -1,4 +1,6 @@
-# coding: utf-8
+
+# frozen_string_literal: true
+
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
 require 'terminal_chess/version'
@@ -14,7 +16,7 @@ Gem::Specification.new do |s|
   s.homepage    = 'https://github.com/at1as/Terminal-Chess'
   s.license     = 'MIT'
 
-  s.files       = `git ls-files`.split("\n") 
+  s.files       = `git ls-files`.split("\n")
   s.require_paths = ["lib"]
   s.executables << 'terminal_chess'
 
@@ -23,6 +25,6 @@ Gem::Specification.new do |s|
   s.add_runtime_dependency "eventmachine"
   s.add_runtime_dependency "faye-websocket"
   s.add_development_dependency "bundler", "~> 1.7"
-  s.add_development_dependency "rake", "~> 10.0"
   s.add_development_dependency "minitest"
+  s.add_development_dependency "rake", "~> 10.0"
 end
