+++ test/helpers.rb
@@ -1,0 +1,8 @@
+module Helpers
+  class MyIO
+    # Send 'Q' to stdin for piece promotion
+    def gets
+      "Q\n"
+    end
+  end
+end
